+++ src/Data/Fields/FieldValueInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Storyblok\ManagementApi\Data\Fields;
+
+interface FieldValueInterface
+{
+    /**
+     * @return array<mixed>
+     */
+    public function toArray(): array;
+}