+++ src/Data/Fields/Schema/FieldNamedConstructor.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Storyblok\ManagementApi\Data\Fields\Schema;
+
+trait FieldNamedConstructor
+{
+    /**
+     * @param mixed[] $data
+     */
+    public static function make(string $key, array $data = []): self
+    {
+        return new self($key, $data);
+    }
+}